eblocks_backend/app/requirements.txt
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
distlib==0.3.4
fastapi==0.78.0
FastAPI-SQLAlchemy==0.2.1
fastapi-utils==0.2.1
filelock==3.7.1
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
keyring==23.6.0
lockfile==0.12.2
Mako==1.2.0
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.1.13
poetry-core==1.0.8
ptyprocess==0.7.0
pydantic==1.9.1
pylev==1.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.0
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
tenacity==8.0.1
tomlkit==0.11.0
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
webencodings==0.5.1
zipp==3.8.0
